separate Model declaration
