+++ app/components/models/index.d.ts
@@ -1,0 +1,6 @@
+import { Vector3 } from "@react-three/fiber"
+
+type modelProps = {
+  src: string
+  position: Vector3
+}
